Time: 0 ms (100.00%) | Memory: 9.2 MB (15.49%) - LeetSync

+++ 7-reverse-integer/reverse-integer.c
@@ -1,5 +1,5 @@
 int reverse(int x){
-    long rev = 0;
+    double rev = 0;
     while (x != 0) {
         int pop = x % 10;
         x /= 10;
